backend/requirements.txt
fastapi==0.105.0
uvicorn[standard]==0.23.2
python-multipart==0.0.6
pydantic==2.9.0
python-dotenv==1.1.0
tensorflow>=2.16,<3.0
numpy>=1.26,<2.0
pillow==10.1.0
librosa==0.10.0
soundfile==0.12.1
pydantic-settings==2.0.0
opencv-python==4.8.1.78
joblib==1.3.2
